+++ SplitWiserClient/src/main/java/com/splitwiser/splitwiserclient/model/payment/Payment.java
@@ -86,9 +86,6 @@ public class Payment {
         return group;
     }
 
-    public BigDecimal getValue() {
-        return amount.get();
-    }
 
     public ObjectProperty<BigDecimal> valueProperty() {
         return amount;